fixed the SW double-tap settings reading updated pt-PT translations
