+++ app/src/main/java/tk/wasdennnoch/androidn_ify/recents/doubletap/DoubleTapSwKeys.java
@@ -26,10 +26,14 @@ public class DoubleTapSwKeys extends DoubleTapBase {
             sOriginalRecentsClickListener.onClick(sRecentsButton);
         }
     };
+
     private static XC_MethodHook prepareNavigationBarViewHook = new XC_MethodHook() {
         @Override
         protected void afterHookedMethod(MethodHookParam param) throws Throwable {
             final Context mContext = (Context) XposedHelpers.getObjectField(param.thisObject, "mContext");
+
+            registerReceiver(mContext);
+
             if (!isTaskLocked(mContext)) {
                 final Handler mHandler = (Handler) XposedHelpers.getObjectField(param.thisObject, "mHandler");
                 sOriginalRecentsClickListener = (View.OnClickListener) XposedHelpers.getObjectField(param.thisObject, "mRecentsClickListener");
@@ -64,7 +68,6 @@ public class DoubleTapSwKeys extends DoubleTapBase {
 
                 Class<?> classPhoneStatusBar = XposedHelpers.findClass(CLASS_PHONE_STATUS_BAR, classLoader);
                 XposedHelpers.findAndHookMethod(classPhoneStatusBar, "prepareNavigationBarView", prepareNavigationBarViewHook);
-
             }
 
         } catch (Throwable t) {
