+++ app/src/main/res/values-pt-rPT/strings.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 
-    <string name="xposed_description">Porta funções introduzidas no Android N a dispositivos com Lollipop e Marshmallow!</string>
+    <string name="xposed_description">Traz funções introduzidas no Android N a dispositivos com Lollipop e Marshmallow!</string>
 
     <string name="header_app">App</string>
 
     <!-- prefs -->
-    <string name="hide_launcher_icon">Pode requerer um reinício do launcher</string>
+    <string name="hide_launcher_icon">Pode ser necessário reiniciar o launcher</string>
     <string name="hide_launcher_icon_title">Esconder ícone do launcher</string>
     <string name="debug_log_title">Registos de depuração</string>
     <string name="debug_log">Atenção: Registo invasivo</string>
@@ -14,13 +14,13 @@
 
 
     <!-- SETTINGS -->
-    <string name="disabled">Desativado</string>
+    <string name="disabled">Desactivado</string>
     <string name="disconnected">Desconectado</string>
-    <string name="off">DESATIVADO</string>
-    <string name="on">ATIVADO</string>
+    <string name="off">DESACTIVADO</string>
+    <string name="on">ACTIVADO</string>
     <string name="data_usage_summary">%1$s de dados utilizados</string>
     <string name="display_summary">Brilho adaptável %1$s</string>
-    <string name="sound_summary">Volume da campainha em %1$s</string>
+    <string name="sound_summary">Volume do toque em %1$s</string>
     <string name="apps_summary">%1$d aplicações instaladas</string>
     <string name="storage_summary">%1$s de %2$s utilizados</string>
     <string name="battery_summary">%1$s - aprox. %2$s restantes</string>
@@ -29,14 +29,21 @@
     <string name="battery_charging_ac">%1$s - A carregar via CA</string>
     <string name="memory_summary">Utilização de memória %1$s de %2$s </string>
     <string name="user_summary">Conectado como %1$s</string>
-    <string name="location_on_power">ATIVADA / Poupança de energia</string>
-    <string name="location_on_device">ATIVADA / Apenas no dispositivo</string>
-    <string name="location_on_high">ATIVADA / Alta precisão</string>
+    <string name="location_on_power">ACTIVADA / Poupança de energia</string>
+    <string name="location_on_device">ACTIVADA / Apenas no dispositivo</string>
+    <string name="location_on_high">ACTIVADA / Alta precisão</string>
     <string name="print_summary">%1$d trabalhos de impressão</string>
     <!-- /SETTINGS -->
 
     <!-- RECENTS -->
     <string name="no_previous_recents">Nenhuma app recente</string>
+    <string name="double_tap_speed_title">Velocidade de duplo toque</string>
+    <string name="enable_recents">Activar modificações em recentes</string>
+    <string name="enable_settings">Activar modificações nas definições</string>
+    <string name="fix_sound_notif_tile">Algumas ROMs juntam o módulo de Sons e Notificações. Activar para mostrar o sumário correcto para estes casos</string>
+    <string name="fix_sound_notif_tile_title"><![CDATA[Corrigir o módulo de "Sons & Notificações"]]></string>
+    <string name="header_recents">Recentes</string>
+    <string name="header_settings">Definições</string>
     <!-- /RECENTS -->
 
 </resources>
